+++ internal/web/public/assets/swimlanesDiagram-VR7AAH4N-akVBx_yH.js
@@ -1,8 +1,0 @@
-import{c as r,s as e}from"./flowDiagram-HODETNUW-CA7VLwjX.js";import{_ as a}from"./mermaid.core-C9frPdAf.js";import"./chunk-5VM5RSS4-eRDAjk4g.js";import"./chunk-XXDRQBXY-DhLr9nWT.js";import"./chunk-POPQ4Y6H-CJhPGFRo.js";import"./chunk-F27PBJKO-BCyA33sn.js";import"./channel-B_r91vt5.js";import"./index-ClFLxTtA.js";var o=a(t=>`${e(t)}
-  .swimlane.cluster rect {
-    stroke: ${t.clusterBorder} !important;
-  }
-  [data-look="neo"].cluster rect {
-    filter: none;
-  }
-`,"getStyles"),s=o,n=r({defaultLayout:"swimlane",styles:s});export{n as diagram};